requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.1
werkzeug==3.1.3
pandas==2.2.3
openpyxl==3.1.5
gunicorn==21.2.0
